Implement cancel events web-animations bonus
Building on dstockwell's web-animations 'bonus files' patch
web-animations/web-animations-next/pull/413, this patch adds
animation cancel events to web-animations and web-animations-next.
The implementation and tests are based on a contribution by
mgiuffrida: web-animations/web-animations-js/pull/58. Whether the
polyfill is loaded is based on whether Animation.oncancel is defined,
and so applies on top of the web-animations polyfill or on top of the
native Element.animate, depending on what is available.

As part of this patch, cancelling an animation no longer causes a
finish event to fire. The test around this case has been removed from
the animation-finish-event.js test suite and added in modified form to
the new animation-cancel-event.js test suite.