Always load scope/shared without guard.
With the addition of webAnimations1BonusSrc files, it is useful for the
web-animations target to load the polyfill scopeSrc and sharedSrc files
whether or not a native web animations implementation exists. This
patch moves the scopeSrc and sharedSrc files outside of the
webAnimations1Src guard, as is done in the web-animations-next target.
This allows the build* functions in Gruntfile.js to be simplified and
merged.