Implement object-form keyframe syntax.
This patch extends the web-animations and web-animations-next
polyfills to accept the object-form syntax for specifying keyframes
(http://w3c.github.io/web-animations/#processing-a-frames-argument).
The capacity is added to the base polyfill, so loading the
web-animations polyfill will enable the new syntax.

It is possible that a browser has a native Web Animations
implementation (and hence does not load the base polyfill), but does
not natively accept object-form keyframes. To address this, this
patch also adds a file to the 'bonus' sources, so the object-form
keyframe syntax can be loaded independently of the base polyfill.