Prevent cubic bezier from infinite-looping
An infinite loop in the cubic bezier function in src/timing-utilities.js was
causing us to skip interfaces/KeyframeEffect/effect-easing.html from the
web-platform-tests. This patch changes the loop condition to ensure it exits,
in line with the CubicBezier implementation in Chromium.

In order to achieve the accuracy required by the web-platform-tests
expectations, the epsilon value in the algorithm is also reduced by an order of
magnitude.