Update README.md with installation instructions
Add installation instructions. The README didn't reference npm even though there is an npm package: #134 had a TODO for adding the installation instructions to the README.

Are these installation instructions right? Or is it `require('web-animation-js').polyfill()` instead of just `require('web-animation-js')`?

Thank you for your work on this library and the cool demos.